refactor(suite): extract CoinsList from CoinsGroup component to be used without CoinsCount in Add Account modal
