+++ packages/suite/src/components/suite/index.tsx
@@ -50,6 +50,7 @@ import TroubleshootingTips from './TroubleshootingTips';
 import ConnectDevicePrompt from './ConnectDevicePrompt';
 import Coin from './Coin';
 import CoinsGroup from './CoinsGroup';
+import CoinsList from './CoinsGroup/CoinsList';
 
 export {
     Backdrop,
@@ -108,5 +109,6 @@ export {
     ConnectDevicePrompt,
     Coin,
     CoinsGroup,
+    CoinsList,
 };
 export type { ImageProps, ModalProps };
